pyproject.toml
[project]
name = "linnaeus"
version = "0.1.0-alpha.1"
description = "Linnaeus: A deep learning framework for taxonomic recognition."
authors = [
    { name = "Caleb Sowers", email = "caleb@polli.ai" },
    # Add other contributors if any
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    # Core ML dependencies
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "yacs>=0.1.8",
    "numpy>=1.20",
    
    # Data processing and utilities
    "tqdm",
    "h5py",
    "wandb",
    "pyyaml",
    "Pillow",
    "opencv-python",
    "pandas",
    "matplotlib",
    "termcolor",
    "rich",
    "polli-typus>=0.1.7",
    "huggingface-hub",
    "python-dateutil"
    
    # Note: flash-attn must be installed separately with --no-build-isolation
    # See setup/install_deps.sh for proper installation order
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "uv",
]

# Flash-attention must be installed separately due to build dependency issues
flash-attn = [
    "flash-attn>=2.5.9.post1",
]

rl = [
    "gymnasium>=0.29.1",
]

[project.urls]
Homepage = "https://github.com/polli-labs/linnaeus"
Issues = "https://github.com/polli-labs/linnaeus/issues"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["linnaeus*"]
exclude = ["work*", "setup*", "archive*", "configs*"]

# ---------------------------------------------------------------------------
# Ruff -- relaxed baseline
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 140                 # allow moderately long research strings
target-version = "py310"
respect-gitignore = true          # skip data/output folders already ignored
include = ["linnaeus/**/*.py"]    # only check files in linnaeus module
exclude = [
    "work/**/*.py",              # exclude work directory
    "tests/**/*.py",             # exclude tests directory
    "tools/**/*.py",             # exclude tools directory
    "setup/**/*.py",             # exclude setup directory
    "archive/**/*.py",           # exclude archive directory
    "configs/**/*.py",           # exclude configs directory
]

[tool.ruff.format]                # formatter runs first; low-stress defaults
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = true  # avoids giant diffs on dict / param lists
line-ending = "auto"

[tool.ruff.lint]
select  = ["E", "F", "W", "UP", "B"]   # core error classes + bugbear safety
extend-select = ["I"]                  # import-sorter (zero-config, low noise)
ignore  = [
  "E501",   # line too long         (handled implicitly by formatter)
  "E741",   # ambiguous variable names (too noisy on legacy code)
  "B008",   # dict/list default values in function defs (acceptable for now)
]
# Temporarily silence whole rule‐groups that dominate the current noise floor
extend-ignore = [
  "ANN",    # missing type-hints
  "C4",     # inconsistent comprehensions
  "TCH",    # typing-only imports
  "ASYNC",  # async-specific suggestions
  "D",      # docstring style/format
]

[tool.ruff.lint.mccabe]
max-complexity = 20               # pragmatic ceiling; revisit once stable

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]  # re-export patterns & late imports are fine
"tests/**/*.py" = ["I", "F401"]   # test util imports & order don't matter yet